requirements.txt
streamlit==1.35.0
requests>=2.31.0
shapely>=2.0.3
pyproj>=3.6.1

# Satellite steps (optional but recommended)
numpy>=1.26.4
pystac-client>=0.7.6
planetary-computer>=1.0.0
odc-stac>=0.3.10
xarray>=2024.1.0
rasterio>=1.3.10
rioxarray>=0.15.5
